requirements.txt
gradio
langchain

pdfminer.six
sentence-transformers
wikipedia